chore: cleanup stuff a bit
